[MOD] 修改id类型 (modify id type to size_t)

+++ src/OE/EasyPool/Task.h
@@ -44,13 +44,13 @@ class OETask
 protected:
 
     // 任务的唯一标识
-    unsigned int id_;
+    size_t id_;
     // 任务创建时间 （非Unix时间戳）
     clock_t createTime_;
 
 private:
 
-    static unsigned int nRequestID_;
+    static size_t nRequestID_;
     // 任务取消状态
     std::atomic<bool>  isCancelRequired_;
 
@@ -74,7 +74,7 @@ public:
     }
 
 	// 获取任务ID
-    unsigned int getID(void){ return id_; }
+    size_t getID(void){ return id_; }
     // 获取任务取消状态
     bool isCancelRequired(){ return isCancelRequired_; }
 	// 设置任务取消状态
@@ -83,6 +83,6 @@ public:
 
 };
 
-__declspec(selectany) unsigned int OETask::nRequestID_ = 100000;
+__declspec(selectany) size_t OETask::nRequestID_ = 100000;
 
 #endif // __OETASK_H__
